# apps/landing/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fluxby API",
    "version": "1.0.0",
    "description": "API for managing financial transactions, categories, budgets, and analytics. This API is intended for developers building custom interfaces.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/fluxby/fluxby/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/budgets": {
      "get": {
        "summary": "Haal alle budgetten op met uitgaven",
        "tags": [
          "Budgets"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "schema": {
              "type": "string"
            },
            "description": "Maand in formaat YYYY-MM (standaard huidige maand)"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string"
            },
            "description": "Start datum in formaat YYYY-MM-DD (optioneel, neemt voorrang over month)"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string"
            },
            "description": "Eind datum in formaat YYYY-MM-DD (optioneel)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lijst met budgetten inclusief besteed bedrag en percentage"
          }
        }
      },
      "post": {
        "summary": "Maak een nieuw budget aan",
        "tags": [
          "Budgets"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "categoryId": {
                    "type": "integer"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "period": {
                    "type": "string",
                    "enum": [
                      "monthly",
                      "yearly"
                    ]
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Budget aangemaakt"
          },
          "400": {
            "description": "Geldig bedrag is verplicht"
          }
        }
      },
      "delete": {
        "summary": "Verwijder alle budgetten",
        "tags": [
          "Budgets"
        ],
        "responses": {
          "200": {
            "description": "Alle budgetten verwijderd"
          }
        }
      }
    },
...
  },
...
}